Add flake.nix
A flake.nix allows everyone to have the same set of packages which
makes developement and distribution easier.

Im adding this because I use nix and already had a flake ready
for a dioxus app, but feel free to remove this and add flake.nix to
the `.gitignore` if you don't want wthis file in the repo.